Move sql transformer in it's own package
