+++ src/test/java/net/datafaker/formats/SqlTest.java
@@ -3,8 +3,8 @@ package net.datafaker.formats;
 import net.datafaker.Faker;
 import net.datafaker.providers.base.BaseFaker;
 import net.datafaker.transformations.Schema;
-import net.datafaker.transformations.SqlDialect;
-import net.datafaker.transformations.SqlTransformer;
+import net.datafaker.transformations.sql.SqlDialect;
+import net.datafaker.transformations.sql.SqlTransformer;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.Arguments;
